% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{choose_discrete_colour_scale}
\alias{choose_discrete_colour_scale}
\title{Chooses a nice colour scale for a discrete number of entries}
\usage{
choose_discrete_colour_scale(number_of_entries)
}
\arguments{
\item{number_of_entries}{input numeric vector (integer) with the number of entries that need colours}
}
\value{
a character vector of colours of the same length as the number provided for the number_of_entries
}
\description{
Will try to choose a colour-blind friendly colour set if possible, but has to move to
less and less ideal colour schemes as the number_of_entries increases
}
