% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_metannotate_table_by_taxon}
\alias{collapse_metannotate_table_by_taxon}
\title{Collapses metannotate table to the given taxonomic rank}
\usage{
collapse_metannotate_table_by_taxon(metannotate_data, taxon = "Family")
}
\arguments{
\item{metannotate_data}{Tibble of metannotate data}

\item{taxon}{Character vector (length 1) giving the taxon name to collapse to
Can be: domain, phylum, class, order, family, genus, species (case insensitive)}
}
\value{
Tibble of metannotate data, collapsed with a 'hits' column
}
\description{
Collapses metannotate table to the given taxonomic rank
}
