% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{create_setup_templates}
\alias{create_setup_templates}
\title{Creates templates for user-specified HMM.Family and Dataset naming data}
\usage{
create_setup_templates(
  metannotate_data,
  write_tables = FALSE,
  hmm_info_filename = "hmm_info_template.tsv",
  dataset_info_filename = "dataset_info_template.tsv"
)
}
\arguments{
\item{metannotate_data}{Tibble of metannotate data}

\item{write_tables}{Logical; write the tables to file?}

\item{hmm_info_filename}{File to optionally write HMM information to}

\item{dataset_info_filename}{File to optionally write dataset information to}
}
\value{
List of two tibbles: hmm_info and dataset_info; user can fill these out
}
\description{
Creates templates for user-specified HMM.Family and Dataset naming data
}
\keyword{internal}
