% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_metannotate_data}
\alias{explore_metannotate_data}
\alias{explore}
\alias{metannotate}
\title{Explore MetAnnotate data}
\usage{
explore_metannotate_data(
  metannotate_data_mapped,
  evalue = 1e-10,
  taxon = "Family",
  normalizing_HMM = "rpoB",
  top_x = 0.02,
  percent_mode = "within_sample",
  colouring_template_filename = NA,
  ...
)
}
\arguments{
\item{metannotate_data_mapped}{The tibble output by \code{\link{map_naming_information}}}

\item{evalue}{E-value cutoff for HMM hits}

\item{taxon}{Character vector (length 1) giving the taxon name to collapse to
Can be: domain, phylum, class, order, family, genus, species (case insensitive)}

\item{normalizing_HMM}{Name of the normalizing HMM (e.g., 'rpoB')]; specify 'auto' to attempt auto-detection}

\item{top_x}{Numeric vector (length 1) giving the subsetting amount you desire.
If top_x >=1, the script will return the "top_x most abundant taxa" for each Dataset/HMM.Family
If top_x <1, the script will return "all taxa of (top_x * 100\%) abundance or greater for each Dataset/HMM.Family -
but see below.}

\item{percent_mode}{If top_x <1, there are two different methods for keeping the most abundant organisms:
\itemize{
\item "within_sample" -- the normalized \% abundance relative to rpoB is used
\item "within_HMM" -- the percent abundance of that taxon within the specific HMM gene hits is used.
You won't notice much of a different between these modes unless one of your HMMs has very few hits and you want to
show some of the taxa that were hit. This would be a good time to use 'within_HMM'.
}}

\item{colouring_template_filename}{Filename of the colouring template you want to load
If the file does not exist, then this function will write a template to that file
If 'NA' is entered, then the function will auto-generate colours and continue on}

\item{...}{Other fine-tuned plotting options controlled by \code{\link{metannotate_plotter}} and the underlying
\code{\link{metannotate_ggplot}}. Highlights include plot_type, which can be "bar" or "bubble"}
}
\value{
A ggplot of MetAnnotate data
}
\description{
high-level exploration function for examining MetAnnotate data
}
