% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_evalue}
\alias{filter_by_evalue}
\alias{filter}
\title{Filter metannotate data by e-value}
\usage{
filter_by_evalue(metannotate_data, evalue = 1e-10)
}
\arguments{
\item{metannotate_data}{Tibble of metannotate data}

\item{evalue}{Numeric vector of the evalue cutoff}
}
\value{
List of two:
\itemize{
\item Tibble of metannotate data, e-value filtered
\item List of three read count tables. First and second are from before and after e-value filtration.
See summarize_total_reads_all_genes(). The third is the \% change between the above two tables.
}
}
\description{
Filters the hits to the desired e-value cutoff and reports stats
}
