% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{generate_plotting_colours}
\alias{generate_plotting_colours}
\title{Generates colour for a metannotate barplot}
\usage{
generate_plotting_colours(metannotate_data_normalized)
}
\arguments{
\item{metannotate_data_normalized}{Tibble of normalized metannotate data - see normalize_collapsed_metannotate_data()}
}
\value{
a tibble of unique taxa with HTML colour codes in the 'colour' column
}
\description{
Generates colour for a metannotate barplot
}
