% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{map_naming_information}
\alias{map_naming_information}
\title{Map user information onto metannotate data}
\usage{
map_naming_information(
  metannotate_data,
  hmm_naming_info_filename,
  dataset_naming_info_filename
)
}
\arguments{
\item{metannotate_data}{Tibble of metannotate data}

\item{hmm_naming_info_filename}{Filepath to HMM naming template exported from \code{\link{create_setup_templates}}}

\item{dataset_naming_info_filename}{Filepath to Dataset naming template exported from
\code{\link{create_setup_templates}}}
}
\value{
Tibble of metannotate data (renamed with factors)
}
\description{
Applies user-provided HMM.Family/Dataset naming/length information onto the metannotate data.
A few interesting quirks about this function:
\itemize{
\item The order of the elements in the provided tables is meaningful. It determines the order of the factor for HMM
Dataset names. This dictates the order during plotting
\item You can give two elements the same plotting name (e.g., R1 and R2 reads). They will be merged together downstream.
\item Only elements in these tables will be used in the script. Any elements you omit from the info tables (e.g., R2 reads,
if you don't want to plot these) will be deleted from the metannotate table.
}
}
