% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{metannotate_ggplot}
\alias{metannotate_ggplot}
\title{Generate a ggplot of the MetAnnotate data}
\usage{
metannotate_ggplot(
  metannotate_data_normalized,
  hit_totals,
  plotting_colour_data,
  plotting_taxon,
  normalizing_HMM,
  plot_type = "bar",
  space = "free",
  bubble_size_range = c(1, 20),
  alpha = 0.8,
  bubble_labels = TRUE
)
}
\arguments{
\item{metannotate_data_normalized}{Tibble of normalized metannotate data - see normalize_collapsed_metannotate_data()}

\item{hit_totals}{Tibble of total normalized hits per HMM/sample
This is generated in normalize_collapsed_metannotate_data()}

\item{plotting_colour_data}{Tibble of plot colours generated by process_plotting_colours()}

\item{plotting_taxon}{The taxonomy rank to summarize bars/bubbles to; MUST MATCH the collapse taxon of the MetAnnotate table}

\item{normalizing_HMM}{the name of the normalizing HMM (e.g., 'rpoB')}

\item{plot_type}{Either 'bar' or 'bubble'}

\item{space}{ggplot setting; 'fixed' or 'free'}

\item{bubble_size_range}{numeric vector of length two with the small and large bubble sizes}

\item{alpha}{ggplot value; transparency (for bubble plots)}

\item{bubble_labels}{logical; show percent labels on the bubbles for bubble plots?}
}
\value{
A ggplot of MetAnnotate data
}
\description{
Generates the ggplot of MetAnnotate data
}
