% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{metannotate_plotter}
\alias{metannotate_plotter}
\title{Wrapper for convenient generation of MetAnnotate ggplot data}
\usage{
metannotate_plotter(
  metannotate_data_normalized_list,
  colouring_template_filename = NA,
  top_x = NA,
  percent_mode = "within_sample",
  normalizing_HMM = "auto",
  plot_normalizing_HMM = TRUE,
  dump_raw_data = FALSE,
  ...
)
}
\arguments{
\item{metannotate_data_normalized_list}{List output of normalize_collapsed_metannotate_data()}

\item{colouring_template_filename}{Filename of the colouring template you want to load
If the file does not exist, then this function will write a template to that file
If 'NA' is entered, then the function will auto-generate colours and continue on}

\item{top_x}{Numeric vector (length 1) giving the subsetting amount you desire.
If top_x >=1, the script will return the "top_x most abundant taxa" for each Dataset/HMM.Family
If top_x <1, the script will return "all taxa of (top_x * 100\%) abundance or greater for each Dataset/HMM.Family - but see below.}

\item{percent_mode}{If top_x <1, there are two different methods for keeping the most abundant organisms:
\itemize{
\item "within_sample" -- the normalized \% abundance relative to rpoB is used
\item "within_HMM" -- the percent abundance of that taxon within the specific HMM gene hits is used.
You won't notice much of a different between these modes unless one of your HMMs has very few hits and you want to
show some of the taxa that were hit. This would be a good time to use 'within_HMM'.
}}

\item{normalizing_HMM}{Name of the normalizing HMM (e.g., 'rpoB')]; specify 'auto' to attempt auto-detection}

\item{plot_normalizing_HMM}{Retain the normalizing_HMM in the final ggplot?}

\item{dump_raw_data}{Return the normalized and subsetted table in lieu of a ggplot}

\item{...}{Other fine-tuned plotting options controlled by metannotate_ggplot()}
}
\value{
A ggplot of MetAnnotate data (or raw data; see above)
}
\description{
Wrapper to generate a ggplot of MetAnnotate data with subset, colours, labels, and so on
}
