% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metannoviz-package.R
\docType{package}
\name{metannoviz-package}
\alias{metannoviz}
\alias{metannoviz-package}
\title{metannoviz: Process and visualize output from the MetAnnotate pipeline}
\description{
This package facilitates downstream analysis after using
    the MetAnnotate pipeline to scan unassembled metagenome reads from
	environmental metagenome samples for functional genes of interest.
	This package allows for quick exploration and visualization of the
	relative abundances of functional genes compared to a reference taxonomic
	marker gene.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmtsuji/metannotate-analysis}
  \item Report bugs at \url{https://github.com/jmtsuji/metannotate-analysis/issues}
}

}
\author{
\strong{Maintainer}: Jackson Tsuji \email{jackson.tsuji@uwaterloo.ca} (\href{https://orcid.org/0000-0001-7244-4709}{ORCID})

}
\keyword{internal}
