% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{process_plotting_colours}
\alias{process_plotting_colours}
\title{Generates colour for a metannotate barplot or loads user-defined colours}
\usage{
process_plotting_colours(
  metannotate_data_normalized,
  colouring_template_filename = NA
)
}
\arguments{
\item{metannotate_data_normalized}{Tibble of normalized metannotate data - see normalize_collapsed_metannotate_data()}

\item{colouring_template_filename}{Filename of the colouring template you want to load
If the file does not exist, then this function will write a template to that file
If 'NA' is entered, then the function will auto-generate colours and continue on}
}
\value{
a tibble of unique taxa with HTML colour codes in the 'colour' column
}
\description{
A wrapper for generate_plotting_colours() to handle bigger-picture decision making
}
