% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{read_metannotate_data}
\alias{read_metannotate_data}
\title{Loads MetAnnotate data table}
\usage{
read_metannotate_data(metannotate_table_filename)
}
\arguments{
\item{metannotate_table_filename}{filepath to the metannotate table (TSV format; can be compressed)}
}
\value{
tibble of metannotate data
}
\description{
Loads MetAnnotate data table
}
