% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{subset_normalized_metannotate_data}
\alias{subset_normalized_metannotate_data}
\alias{subset}
\title{Subset high-abundance taxa from metannotate data}
\usage{
subset_normalized_metannotate_data(
  metannotate_data_normalized,
  top_x,
  percent_mode = "within_sample"
)
}
\arguments{
\item{metannotate_data_normalized}{Tibble of normalized metannotate data - see
\code{\link{normalize_collapsed_metannotate_data}}}

\item{top_x}{Numeric vector (length 1) giving the subsetting amount you desire.
If top_x >=1, the script will return the "top_x most abundant taxa" for each Dataset/HMM.Family
If top_x <1, the script will return "all taxa of (top_x * 100\%) abundance or greater for each Dataset/HMM.Family -
but see below.}

\item{percent_mode}{If top_x <1, there are two different methods for keeping the most abundant organisms:
\itemize{
\item "within_sample" -- the normalized \% abundance relative to rpoB is used
\item "within_HMM" -- the percent abundance of that taxon within the specific HMM gene hits is used.
You won't notice much of a different between these modes unless one of your HMMs has very few hits and you want to
show some of the taxa that were hit. This would be a good time to use 'within_HMM'.
}}
}
\value{
List of two:
\itemize{
\item Tibble of metannotate data; 'hits' has now been changed to 'percent_abundance' and is double-normalized
\item Tibble summarizing total normalized counts for genes
}
}
\description{
Subsets normalized metannotate data to some desired threshold of most abundance organisms
}
