% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{summarize_total_reads}
\alias{summarize_total_reads}
\title{Counts all HMM hits for the gene of interest}
\usage{
summarize_total_reads(metannotate_data, gene = "rpoB", collapsed = FALSE)
}
\arguments{
\item{metannotate_data}{Tibble of metannotate data}

\item{gene}{Character vector (length 1) of a gene to summarize}

\item{collapsed}{Logical (length 1) - has the table already been collapsed to a taxonomy rank?
If TRUE, then the function will sum the "hits" column instead of a simple count.}
}
\value{
Tibble of summarized hit counts
}
\description{
Counts all HMM hits for the gene of interest
}
